python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
edge-tts>=6.1.9
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
Pillow>=10.2.0
psutil>=5.9.8
pytrends>=4.9.2
